move RoleAssignmentService to the service package
point ref data env var to local ia-docker by default and
point it to the mini-kube wiremock in the application-local.yaml
